Add gradle-based build of all-in-one JAR.
In addition to using gradle build script to create an all-in-one JAR for
the oltpbenchmark, it also creates a bash script that uses this JAR.
The gradle build script leverages the existing ant build script.

To build everything, issue "gradlew build" in the project main drectory
and checkout the output in the build subdirectory.

The commit includes gradle wrapper runtim. Using the gradle wrapper
eliminates the need for having gradle installed on the build host
machine; the wrapper pulls in necessary files when run.